+++ devbox/apps/index.js
@@ -12,7 +12,6 @@ export { default as Checkbox } from './Checkbox'
 export { default as DataView } from './DataView'
 export { default as DiscButton } from './DiscButton'
 export { default as DropDown } from './DropDown'
-export { default as _DropDown } from './_DropDown'
 export { default as EmptyStateCard } from './EmptyStateCard'
 export { default as EthIdenticon } from './EthIdenticon'
 export { default as FloatIndicator } from './FloatIndicator'
