+++ src/components/DropDown/DropDown.js
@@ -1,164 +1,230 @@
-import React from 'react'
+import React, { useState, useRef, useEffect, useCallback } from 'react'
 import PropTypes from 'prop-types'
-import styled from 'styled-components'
-import { Transition, animated } from 'react-spring'
-import ClickOutHandler from 'react-onclickout'
-import { theme } from '../../theme-legacy'
-import { springs } from '../../style'
+import { Button, ButtonBase, Popover } from '../'
+import { IconDown } from '../../icons'
+import { GU, RADIUS, textStyle } from '../../style'
+import { useTheme } from '../../theme'
 import { unselectable } from '../../utils'
-import { PublicUrl } from '../../providers/PublicUrl'
-import DropDownItem from './DropDownItem'
 
-import arrow from './assets/arrow-down.svg'
+function useDropDown({ items, selected, onChange, label }) {
+  const containerRef = useRef()
+  const [selectedLabel, setSelectedLabel] = useState(label)
+  const [opened, setOpened] = useState(false)
+  const handleClose = useCallback(() => {
+    // if the popover is opened and the user clicks on the button
+    // this handler was being called before the click handler, so the
+    // click handler was re-opening the popover, by having this on the
+    // next tick things happen in order.
+    setTimeout(() => setOpened(false), 10)
+  }, [setOpened])
+  const handleToggle = useCallback(() => setOpened(!opened), [
+    setOpened,
+    opened,
+  ])
+  const handleChange = useCallback(
+    index => {
+      onChange(index)
+      handleClose()
+      setSelectedLabel(items[index])
+    },
+    [onChange, handleClose, items, setSelectedLabel]
+  )
 
-const NON_BREAKING_SPACE = '\xa0'
-
-const { contentBackground, contentBorder, textPrimary } = theme
-
-const StyledDropDown = styled.div`
-  position: relative;
-  z-index: ${({ opened }) => (opened ? '2' : '0')};
-  display: ${({ wide }) => (wide ? 'flex' : 'inline-flex')};
-  flex-direction: column;
-  color: ${textPrimary};
-  white-space: nowrap;
-  box-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.03);
-  ${unselectable()};
-  &:focus {
-    outline: 0;
-  }
-`
-
-const DropDownItems = styled(animated.div)`
-  position: absolute;
-  z-index: 2;
-  top: calc(100% - 1px);
-  padding: 8px 0;
-  color: ${textPrimary};
-  background: ${contentBackground};
-  border: 1px solid ${contentBorder};
-  box-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.06);
-  border-radius: 3px;
-  list-style: none;
-`
-
-const BlockingLayer = styled(animated.div)`
-  position: absolute;
-  z-index: 2;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-`
-
-const DropDownActiveItem = styled(PublicUrl.hocWrap(DropDownItem))`
-  padding-right: 40px;
-  background: ${contentBackground};
-  background-image: url(${PublicUrl.styledUrl(arrow)});
-  background-repeat: no-repeat;
-  background-position: calc(100% - 15px) 50%;
-  border: 1px solid ${contentBorder};
-  border-radius: 3px;
-  &:hover,
-  &:focus {
-    color: inherit;
-  }
-  &:active {
-    color: ${textPrimary};
-  }
-`
-
-class DropDown extends React.Component {
-  static propTypes = {
-    items: PropTypes.arrayOf(PropTypes.node),
-    wide: PropTypes.bool,
-    active: PropTypes.number,
-    onChange: PropTypes.func,
-  }
-  static defaultProps = {
-    items: [],
-    wide: false,
-    active: 0,
-    onChange: () => {},
-  }
-  state = {
-    opened: false,
-  }
-  activeItemElt = null
-
-  handleToggle = () => {
-    this.setState({ opened: !this.state.opened })
-  }
-  handleClose = () => {
-    this.setState({ opened: false })
-  }
-  handleItemActivate = (index, { keyboard }) => {
-    this.props.onChange(index, this.props.items)
-    this.setState({ opened: false })
-    if (this.activeItemElt && keyboard) {
-      this.activeItemElt.focus()
+  useEffect(() => {
+    if (selected === -1 || !items[selected]) {
+      return
     }
+    setSelectedLabel(items[selected])
+  }, [items, selected, setSelectedLabel])
+
+  return {
+    containerRef,
+    handleChange,
+    handleClose,
+    handleToggle,
+    opened,
+    selectedLabel,
   }
-  render() {
-    const { items, active, wide } = this.props
-    const { opened } = this.state
-    const activeItem = items[active] || items[0]
-    return (
-      <ClickOutHandler onClickOut={this.handleClose}>
-        <StyledDropDown wide={wide} opened={opened}>
-          <DropDownActiveItem
-            onActivate={this.handleToggle}
-            mainRef={el => (this.activeItemElt = el)}
-          >
-            {activeItem}
-          </DropDownActiveItem>
-          <Transition
-            config={springs.swift}
-            items={opened}
-            from={{ scale: 0.98, opacity: 0, enabled: 1 }}
-            enter={{ scale: 1, opacity: 1, enabled: 1 }}
-            leave={{ scale: 1, opacity: 0, enabled: 0 }}
-            native
-          >
-            {opened =>
-              opened &&
-              (({ scale, opacity, enabled }) => (
-                <DropDownItems
-                  role="listbox"
-                  style={{
-                    opacity,
-                    transform: scale.interpolate(t => `scale3d(${t},${t},1)`),
-                    minWidth: wide ? '100%' : '0',
-                  }}
-                >
-                  {items.length
-                    ? items.map((item, i) => (
-                        <DropDownItem
-                          role="option"
-                          key={i}
-                          index={i}
-                          active={i === active}
-                          onActivate={this.handleItemActivate}
-                        >
-                          {item}
-                        </DropDownItem>
-                      ))
-                    : NON_BREAKING_SPACE}
-                  <BlockingLayer
-                    style={{
-                      display: enabled.interpolate(t =>
-                        t === 1 ? 'none' : 'block'
-                      ),
-                    }}
-                  />
-                </DropDownItems>
-              ))
-            }
-          </Transition>
-        </StyledDropDown>
-      </ClickOutHandler>
+}
+
+const DropDown = React.memo(function DropDown({
+  active,
+  header,
+  items,
+  label,
+  onChange,
+  selected,
+  width,
+}) {
+  if (active !== undefined) {
+    console.warn(
+      'The “active” prop is deprecated. Please use “selected” to pass the selected index instead.'
     )
   }
+  const selectedIndex = active || selected
+  const theme = useTheme()
+  const {
+    containerRef,
+    handleChange,
+    handleClose,
+    handleToggle,
+    opened,
+    selectedLabel,
+  } = useDropDown({
+    selectedIndex,
+    items,
+    label,
+    onChange,
+  })
+  const closedWithChanges = !opened && selectedIndex !== -1
+
+  return (
+    <div ref={containerRef}>
+      <Button
+        onClick={handleToggle}
+        css={`
+          display: flex;
+          justify-content: space-between;
+          align-items: center;
+          padding: 0 ${2 * GU}px;
+          width: ${width};
+          min-width: unset;
+          ${closedWithChanges ? `border: 1px solid ${theme.accent}` : ''}
+        `}
+      >
+        <span
+          css={`
+            color: ${theme.content};
+            ${textStyle('body2')};
+          `}
+        >
+          {selectedLabel}
+        </span>
+        <IconDown
+          size="tiny"
+          css={`
+            transition: transform 0.3s;
+            transform: rotate3d(0, 0, 1, ${opened ? 180 : 0}deg);
+            ${closedWithChanges ? `color: ${theme.accent}` : ''}
+          `}
+        />
+      </Button>
+      <Popover
+        visible={opened}
+        onClose={handleClose}
+        opener={containerRef.current}
+        placement="bottom-start"
+      >
+        <div
+          css={`
+            width: calc(${width} - 2px);
+            color: ${theme.surfaceContentSecondary};
+          `}
+        >
+          {header && (
+            <div
+              css={`
+                padding: ${1.5 * GU}px ${2 * GU}px ${1.25 * GU}px;
+                ${textStyle('label2')};
+                ${unselectable};
+              `}
+            >
+              {header}
+            </div>
+          )}
+          <ul
+            css={`
+              margin: 0;
+              list-style: none;
+              width: 100%;
+            `}
+          >
+            {items.map((item, index) => {
+              return (
+                <Item
+                  key={index}
+                  index={index}
+                  onClick={handleChange}
+                  theme={theme}
+                  item={item}
+                  header={header}
+                  length={items.length}
+                  selected={selectedIndex}
+                />
+              )
+            })}
+          </ul>
+        </div>
+      </Popover>
+    </div>
+  )
+})
+
+DropDown.propTypes = {
+  active: PropTypes.number,
+  header: PropTypes.node,
+  items: PropTypes.arrayOf(PropTypes.node).isRequired,
+  label: PropTypes.string,
+  onChange: PropTypes.func.isRequired,
+  selected: PropTypes.number.isRequired,
+  width: PropTypes.string,
+}
+
+DropDown.defaultProps = {
+  label: 'Select an item',
+}
+
+const Item = React.memo(function Item({
+  header,
+  index,
+  item,
+  length,
+  onClick,
+  selected,
+  theme,
+}) {
+  const handleClick = useCallback(() => {
+    onClick(index)
+  }, [index, onClick])
+
+  return (
+    <li>
+      <ButtonBase
+        onClick={handleClick}
+        css={`
+          width: 100%;
+          text-align: left;
+          padding: ${1.25 * GU}px ${2 * GU}px;
+          border-radius: 0;
+          color: ${theme.content};
+          ${textStyle('body2')};
+          ${!header && index === 0 && `border-top-left-radius: ${RADIUS}px;`}
+          ${index === length - 1 && `border-bottom-left-radius: ${RADIUS}px;`}
+          ${selected === index &&
+            `
+              border-left: 2px solid ${theme.accent};
+              background: ${theme.surfaceSelected};
+            `}
+
+          &:active {
+            background: ${theme.surfacePressed};
+          }
+        `}
+      >
+        {item}
+      </ButtonBase>
+    </li>
+  )
+})
+
+Item.propTypes = {
+  header: PropTypes.node,
+  index: PropTypes.number.isRequired,
+  item: PropTypes.node.isRequired,
+  length: PropTypes.number.isRequired,
+  onClick: PropTypes.func.isRequired,
+  selected: PropTypes.number.isRequired,
+  theme: PropTypes.object.isRequired,
 }
 
 export default DropDown
