+++ src/components/index.js
@@ -23,7 +23,6 @@ export { default as ContextMenu } from './ContextMenu/ContextMenu'
 export { default as ContextMenuItem } from './ContextMenu/ContextMenuItem'
 export { default as Countdown } from './Countdown/Countdown'
 export { default as DropDown } from './DropDown/DropDown'
-export { default as _DropDown } from './DropDown/_DropDown'
 export { default as EmptyStateCard } from './Card/EmptyStateCard'
 export { default as EscapeOutside } from './EscapeOutside/EscapeOutside'
 export { default as EthIdenticon } from './EthIdenticon/EthIdenticon'
